feat comment compoent 분리
