+++ src/components/feed/Feed.jsx
@@ -1,66 +1,29 @@
-import React, { useState, useRef } from "react";
-import { Link } from "react-router-dom";
+import React, { useState } from 'react';
+import { Link } from 'react-router-dom';
 import {
   AiOutlineEllipsis,
   AiOutlineHeart,
   AiOutlineMessage,
   AiOutlineSend,
   AiOutlinePushpin,
-  AiOutlineSmile,
-} from "react-icons/ai";
+} from 'react-icons/ai';
+import CommentWrite from './comment/CommentWrite';
+import CommentList from './comment/CommentList';
 
 const Feed = ({ feed }) => {
-  //이미지 로딩 상태
+  // 이미지 로딩 상태
   const [loaded, setLoaded] = useState(false);
-  //새로 추가한 댓글 comment 담는 곳
-  const [comment, setComment] = useState();
-  //feedData.json파일 댓글
+
+  // feedData.json파일 댓글
   const [comments, setComments] = useState(feed.comments);
 
-  //댓글 input ref
-  const inputComment = useRef();
-
-  //이미지 로딩 함수
+  // 이미지 로딩 함수
   const onLoaded = () => {
     setLoaded(true);
   };
 
-  //댓글 입력 이벤트
-  const onChangeComment = (e) => {
-    const { value } = e.target;
-
-    setComment(value);
-  };
-
-  //댓글 게시 버튼 클릭 이벤트
-  const onPostClickHandler = (e) => {
-    setCommentFunc();
-  };
-
-  //엔터키 입력 이벤트
-  const onPostPressHandler = (e) => {
-    if (e.key === "Enter" && e.nativeEvent.isComposing === false) {
-      setCommentFunc();
-    }
-  };
-
-  //댓글추가 => 기존 데이터랑 합침
-  function setCommentFunc() {
-    //user 이름은 로컬스토리지에서 받아옴.
-    setComments([
-      ...comments,
-      {
-        user: localStorage.getItem("user"),
-        content: comment,
-      },
-    ]);
-
-    //input 초기화
-    inputComment.current.value = "";
-  }
-
   return (
-    <div className={"feed-item " + (loaded && "on")}>
+    <div className={'feed-item ' + (loaded && 'on')}>
       {loaded && (
         <div className="feed-info">
           <div className="user">
@@ -106,34 +69,12 @@ const Feed = ({ feed }) => {
               <p>{feed.content}</p>
             </div>
             {/* 댓글 */}
-            <div className="feed-comment_list">
-              <ul>
-                {comments.map((v, idx) => (
-                  <li key={idx}>
-                    <p>
-                      <strong>{v.user}</strong>
-                      {v.content}
-                    </p>
-                  </li>
-                ))}
-              </ul>
-            </div>
+            <CommentList comments={comments} />
           </div>
-          <div className="feed-comment_input">
-            <div className="input">
-              <AiOutlineSmile />
-              <input
-                type="text"
-                className="w100"
-                ref={inputComment}
-                onChange={onChangeComment}
-                onKeyDown={onPostPressHandler}
-              />
-            </div>
-            <button type="button" onClick={onPostClickHandler}>
-              게시
-            </button>
-          </div>
+          <CommentWrite
+            setComments={setComments}
+            comments={comments}
+          />
         </>
       )}
     </div>
